--- src/components/Commons/ScrollTopButton/index.tsx
@@ -1,9 +1,0 @@
-import { VFC } from 'react';
-
-import { Presenter } from './presenter';
-import { useHooks } from './hooks';
-
-export const ScrollTopButton: VFC = () => {
-  const viewProps = useHooks();
-  return <Presenter {...viewProps} />;
-};
